// global/interfaces/ishop.ts
export interface IRowShop {
	Row1: false;
	Row2: false;
	Row3: false;
	Row4: false;
	Row5: false;
}

export interface IShop {
	Col1: IRowShop;
	Col2: IRowShop;
	Col3: IRowShop;
	Col4: IRowShop;
	Col5: IRowShop;
	Col6: IRowShop;
	Col7: IRowShop;
	Col8: IRowShop;
	Col9: IRowShop;
	Col10: IRowShop;
	Col11: IRowShop;
	Col22: IRowShop;
	Col23: IRowShop;
	Col24: IRowShop;
	Col25: IRowShop;
	Col26: IRowShop;
	Col27: IRowShop;
	Col28: IRowShop;
	Col29: IRowShop;
	Col30: IRowShop;
}